Added custom validation for BusinessName and Categories

+++ CICMain/CICLatest/Controllers/ViewForm1Controller.cs
@@ -110,7 +110,7 @@ namespace CICLatest.Controllers
                 model.WitnessedName = (string)myJObject["value"][i]["WitnessedName"];
                 model.WitnessedTitle = (string)myJObject["value"][i]["WitnessedTitle"];
                 model.Title = (string)myJObject["value"][i]["Title"];
-                model.FirmRegistrationNo = (int)myJObject["value"][i]["FirmRegistrationNo"];
+                model.FirmRegistrationNo = (Int64)myJObject["value"][i]["FirmRegistrationNo"];
                 model.path = (string)myJObject["value"][i]["path"];
                 model.Reviewer = (string)myJObject["value"][i]["Reviewer"];
                 model.Grade = (string)myJObject["value"][i]["Grade"];
